Moved charges.py to openeye.py
